Coded up methods to read the camera and position of white strip
Much of the details of this update is in the comments in the code, but note: this code hasn't been tested or compiled yet as I was on my laptop and wrote it up on github in the vichub. It should if its working  correctly find where in the image the white strip is by defining endpoints yo.